# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presences API",
        "description": "Cara login:\r\n * Login ke endpoint /auth/login dgn email dan password Anda\r\n * Kopi token dari resultnya\r\n * Klik button authorize di pojok kanan atas\r\n * Masukkan token, dengan format::  Bearer <api_key>\r\n * Klik submit button Authorize",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "LIVE server"
        }
    ],
    "paths": {
...
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Retrieve employee list",
                "description": "Retrieve employee list",
                "operationId": "9e1329829c4de5cbb565de40cd6f7721",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "10"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "nip",
                                "name"
                            ]
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "employee list"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Retrieve employee list",
                "description": "Retrieve employees detail",
                "operationId": "38fa0788c8731bbb0f3f7d9c016de542",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "employee detail"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/employees/{id}/worktime": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Add Worktime",
                "description": "Employee add worktime",
                "operationId": "c87ac365d20a9cffc3722af77d0d765a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmployeeWorktime"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "add worktime"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "delete Worktime",
                "description": "Employee delete worktime",
                "operationId": "01e483ca44ea37babde8a183b086a6b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorktimeId"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "delete worktime"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/worktimes": {
            "get": {
                "tags": [
                    "Worktimes"
                ],
                "summary": "Retrieve worktime list",
                "description": "Retrieve worktimes list",
                "operationId": "2c52c5d032843258ff251813c2a41457",
                "responses": {
                    "200": {
                        "description": "worktime list"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Worktimes"
                ],
                "summary": "Create worktimes",
                "description": "Create worktimes",
                "operationId": "6f09341620c3c4159deb6f361c995397",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Worktime"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Worktime created"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/worktimes/{id}": {
            "get": {
                "tags": [
                    "Worktimes"
                ],
                "summary": "Retrieve worktime list",
                "description": "Retrieve worktimes detail",
                "operationId": "8c5c7da584b0fd01958a446d4268cb9b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Worktime detail"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Worktimes"
                ],
                "summary": "Update worktimes",
                "description": "Update worktimes",
                "operationId": "e8990c93563b78e5baefa9a32cca2102",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Worktime"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Worktime updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Worktimes"
                ],
                "summary": "Delete worktimes",
                "description": "Delete worktimes",
                "operationId": "02e72a3e60a6b589ef8ef6bda0d8aca4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Worktime deleted"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/worktimes/{id}/workunit": {
            "post": {
                "tags": [
                    "Worktimes"
                ],
                "summary": "Assign workunits on worktime",
                "description": "Assgin workunits",
                "operationId": "d2b40c7a74784e21ebd6ddfbdb4e2bd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorktimeWorkunits"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Workunit assigned"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmployeeWorktime": {
                "properties": {
                    "worktime_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "date_start": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-02-01"
                    }
                },
                "type": "object"
            },
...
            "Worktime": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Example Name"
                    }
                },
                "type": "object"
            },
            "WorktimeId": {
                "properties": {
                    "worktime_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "WorktimeWorkunits": {
                "properties": {
                    "workunit": {
                        "type": "array",
                        "items": "integer",
                        "example": "[1,2,3]"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}